[mem] add mem_rep_copy()
There are two parameters that may need tweaking: the effect upon the end of the
destination slice if it is not large enough for the last stride, and the effect
of a zero stride.  The correct choices for what to do in these situations will
depend upon how this function is used -- at the moment, the safest options have
been chosen.