[mem] overhaul memory comparisons
mem_islt(), mem_isle(), mem_isgt(), mem_isge(), and mem_coll() have been added,
for performing inequality comparisons between slices of memory.  mem_coll() is
an objectively terrible name for the function, but it has been chosen explicitly
so that the function is difficult to understand, know of, and use -- in almost
all cases, the mem_isXX() comparison functions will suffice.

It also appears that the implementations of mem_iseq() and mem_isne() used bad
parameter names ('src' and 'dst' instead of 'lhs' and 'rhs'); this is fixed.
Furthermore, all boolean comparison functions are implemented in the same source
file, as they will have similar, if not identical, implementations.