rearchitecture, again
The new architecture supports separating platform-specific header files
from platform-independent implementation header files (the difference
is in the APIs exposed).  It is a bit more complex, and relies more on
tup to do the right thing, but it works well.

Note that the types.h definition of 'bool' has been reverted; the goal
is to support C17 primarily, with C23 support as it stabilizes, but in
any case, C23 does _not_ define 'bool', 'true', and 'false' directly.