fix in-tree builds
The !tup_preserve macro rule does nothing in the case of in-tree builds
but the project was relying upon it to copy/link files in order to make
configuration selections.  This has been fixed by using a custom macro
instead that manually symlinks, so that it runs in-tree as well.