add module 'err'
This is a first draft of the error handling module.  I have taken the
dangerous step of writing a full module overview and rationale which
explains how the error handling paradigm expressed here; if this does
not turn out to be what we need, I will have to rewrite it and I will
not enjoy myself.